core code review corrections - non-functional.
